% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceglm.R
\name{niceglm}
\alias{niceglm}
\title{Lisa's GLM Regression Table Reporting Function}
\usage{
niceglm(fit = NA, df = NA, family = NA, covs = NA, out = NA,
  regtype = "multi", exp = NA, estname = NA, intercept = FALSE,
  refcat = FALSE, labels = NA, overallp = FALSE, est.dec = 2,
  ci.dec = 2, pval.dec = 3, color = "#EEEEEE", htmlTable = TRUE,
  title = "", footer = TRUE)
}
\arguments{
\item{fit}{GLM model object [fit or family/covs/out are REQUIRED].}

\item{df}{Dataframe [fit or df/family/covs/out].}

\item{family}{Character. Model family name in quotes ("guassian", "binomial", "poisson") [fit or df/family/covs/out are REQUIRED].}

\item{covs}{Character. Vector of covariates to include in model [fit or df/family/covs/out are REQUIRED].}

\item{out}{Character. Outcome for regression model [fit or df/family/covs/out are REQUIRED].}

\item{regtype}{Logical. Should the covariates be run separately ("uni") or together in a multiple regression model ("multi") [REQUIRED if no fit].}

\item{exp}{Logical. Option to exponentiate coefficients and CI's. Default is NA (estimates exponentiated for binomial and poisson models by default).}

\item{estname}{Character. Label for regression estimate. Default is NA (program will choose a reasonable name depending on model type).}

\item{intercept}{Logical. If TRUE the intercept will be included in the table (muliple regression only). Default is FALSE.}

\item{refcat}{Logical. If TRUE the table will create a separate line for the reference category. Default is FALSE.}

\item{overallp}{Logical. If TRUE, a likelihood ratio test pvalue will be calculated for each variable (via drop1, Chisq test). Default is FALSE.}

\item{est.dec}{Numeric. Number of decimal places for estimates. Default is 2.}

\item{ci.dec}{Numeric. Number of decimal places for 95 percent confidence interval. Default is 2.}

\item{pval.dec}{Numeric. Number of decimal places for pvalues. Must be an integer from 1 to 4. Default is 3.}

\item{color}{Character. Color to use for htmlTable striping. Default is "#EEEEEE" (light grey). Use "white" for no striping.}

\item{htmlTable}{Logical. If TRUE, the table will be optimized for html output. Default is TRUE.}

\item{title}{Character. Optional title above table. Default is "".}

\item{footer}{Logical. If TRUE, table will include a footnote with model details, nobs, R2. Default is TRUE.}
}
\description{
This function creates a nice looking table of regression results for a glm model.
Input either a glm object or dataframe, outcome variable, vector of covariates, model family, and type of analysis (univariate or multiple regression).
The function returns a dataframe of formatted regression results and prints the results table using htmlTable.
}
\keyword{glm}
\keyword{linear}
\keyword{logistic}
\keyword{poisson}
\keyword{regression}
\keyword{reporting}
\keyword{table}
